move: DatePicker Extension Register뷰에서 분리
